Move `Response` class
